Avoid piping to/from S3
This creates a very long connection to S3 which is more likely to get
dropped/cancelled. Fargate disks are now encrypted by default so there
is less concern about copying the file there. It gets shredded after use
for an additional safety measure. The default Fargate disk is 20GB.
We'll need adjust the disk size for bigger dump files in the client.